requirements.txt
Keras==2.4.3
Flask==1.1.2
matplotlib==3.3.2
numpy==1.19.5
tensorflow==2.4.1
gunicorn==20.0.4
flask_cors==3.0.10
Pillow==8.1.2
opencv-python==4.4.0.42
